+++ test/defaults.js
@@ -18,3 +18,13 @@ test('mergeDefaults({ skip: false }) => skip=false', (t) => {
   const options = mergeDefaults({ skip: false });
   t.is(options.skip, false);
 });
+
+test('mergeDefaults({prune: true}) => prune=true', (t) => {
+  const options = mergeDefaults({ prune: true });
+  t.is(options.prune, true);
+});
+
+test('mergeDefaults({prune: false}) => prune=false', (t) => {
+  const options = mergeDefaults({ prune: false });
+  t.is(options.prune, false);
+});